separacion por ventanas v1

+++ main.py
@@ -1,31 +1,41 @@
 import tkinter as tk
 from model.huffman import Huffman
 from model.shannon_fano import ShannonFano
-
-
 from model.utils import Utils
 from view.main_view import MainView
 from controller.main_controller import MainController
 import os
 
-
-
-
 if __name__ == "__main__":
+    # Crear la carpeta temporal si no existe
     temp_image_dir = "temp_images"
     os.makedirs(temp_image_dir, exist_ok=True)
 
     root = tk.Tk()
 
+    try:
 
-    #  modelos
+        current_dir = os.path.dirname(os.path.abspath(__file__))
+        azure_theme_path = os.path.join(current_dir, "azure.tcl")
+
+
+        # Theme Azure: https://github.com/rdbende/Azure-ttk-theme/blob/main/azure.tcl
+
+
+        root.tk.call("source", azure_theme_path)  # Cargar el archivo TCL
+        root.tk.call("set_theme", "light")
+    except Exception as e:
+        print(f"Advertencia: No se pudo cargar el tema. La interfaz usará el tema por defecto. Error: {e}")
+
+
+
+
+    # modelos
     huffman_model = Huffman()
     shannon_fano_model = ShannonFano()
     utils_model = Utils()
 
-
-
-    #  vista
+    # vista principal que gestiona las pantallas
     view = MainView(root)
 
     # controlador
@@ -41,14 +51,13 @@ if __name__ == "__main__":
                 print(f"Error al eliminar archivo temporal: {e}")
 
         try:
-            if not os.listdir(temp_image_dir):
+            if os.path.exists(temp_image_dir) and not os.listdir(temp_image_dir):
                 os.rmdir(temp_image_dir)
         except OSError as e:
             print(f"Error al eliminar directorio temporal: {e}")
         root.destroy()
 
     root.protocol("WM_DELETE_WINDOW", on_closing)
-
